# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResearchRequest {
  /** The investigation target or topic */
  query: string;
}

export type ResearchResultSourcesItem = {
  title?: string;
  uri?: string;
};

export interface ResearchResult {
  query: string;
  content: string;
  sources: ResearchResultSourcesItem[];
}

export type ScriptStatus = (typeof ScriptStatus)[keyof typeof ScriptStatus];

export const ScriptStatus = {
  draft: "draft",
  "in-progress": "in-progress",
  final: "final",
} as const;

export interface Script {
  id: number;
  title: string;
  episode?: string;
  content: string;
  status: ScriptStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateScriptRequestStatus =
  (typeof CreateScriptRequestStatus)[keyof typeof CreateScriptRequestStatus];

export const CreateScriptRequestStatus = {
  draft: "draft",
  "in-progress": "in-progress",
  final: "final",
} as const;

export interface CreateScriptRequest {
  title: string;
  episode?: string;
  content: string;
  status?: CreateScriptRequestStatus;
}

export type UpdateScriptRequestStatus =
  (typeof UpdateScriptRequestStatus)[keyof typeof UpdateScriptRequestStatus];

export const UpdateScriptRequestStatus = {
  draft: "draft",
  "in-progress": "in-progress",
  final: "final",
} as const;

export interface UpdateScriptRequest {
  title?: string;
  episode?: string;
  content?: string;
  status?: UpdateScriptRequestStatus;
}

export type ClipPlatform = (typeof ClipPlatform)[keyof typeof ClipPlatform];

export const ClipPlatform = {
  youtube: "youtube",
  cspan: "cspan",
  other: "other",
} as const;

export interface Clip {
  id: number;
  title: string;
  sourceUrl?: string;
  platform: ClipPlatform;
  timestamp?: string;
  keywords: string[];
  transcript?: string;
  season: string;
  notes?: string;
  createdAt: string;
}

export type CreateClipRequestPlatform =
  (typeof CreateClipRequestPlatform)[keyof typeof CreateClipRequestPlatform];

export const CreateClipRequestPlatform = {
  youtube: "youtube",
  cspan: "cspan",
  other: "other",
} as const;

export interface CreateClipRequest {
  title: string;
  sourceUrl?: string;
  platform: CreateClipRequestPlatform;
  timestamp?: string;
  keywords: string[];
  transcript?: string;
  season: string;
  notes?: string;
}

export type EpisodeStatus = (typeof EpisodeStatus)[keyof typeof EpisodeStatus];

export const EpisodeStatus = {
  planning: "planning",
  scripting: "scripting",
  filming: "filming",
  editing: "editing",
  published: "published",
} as const;

export interface Episode {
  id: number;
  title: string;
  episodeNumber?: number;
  season: string;
  status: EpisodeStatus;
  scheduledDate?: string;
  publishedDate?: string;
  youtubeUrl?: string;
  topic?: string;
  notes?: string;
  createdAt: string;
}

export type CreateEpisodeRequestStatus =
  (typeof CreateEpisodeRequestStatus)[keyof typeof CreateEpisodeRequestStatus];

export const CreateEpisodeRequestStatus = {
  planning: "planning",
  scripting: "scripting",
  filming: "filming",
  editing: "editing",
  published: "published",
} as const;

export interface CreateEpisodeRequest {
  title: string;
  episodeNumber?: number;
  season: string;
  status?: CreateEpisodeRequestStatus;
  scheduledDate?: string;
  topic?: string;
  notes?: string;
}

export type UpdateEpisodeRequestStatus =
  (typeof UpdateEpisodeRequestStatus)[keyof typeof UpdateEpisodeRequestStatus];

export const UpdateEpisodeRequestStatus = {
  planning: "planning",
  scripting: "scripting",
  filming: "filming",
  editing: "editing",
  published: "published",
} as const;

export interface UpdateEpisodeRequest {
  title?: string;
  episodeNumber?: number;
  season?: string;
  status?: UpdateEpisodeRequestStatus;
  scheduledDate?: string;
  publishedDate?: string;
  youtubeUrl?: string;
  topic?: string;
  notes?: string;
}

export type TrendVolume = (typeof TrendVolume)[keyof typeof TrendVolume];

export const TrendVolume = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface Trend {
  id: number;
  title: string;
  volume: TrendVolume;
  sentiment: string;
  active: boolean;
  createdAt: string;
}

export type CreateTrendRequestVolume =
  (typeof CreateTrendRequestVolume)[keyof typeof CreateTrendRequestVolume];

export const CreateTrendRequestVolume = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface CreateTrendRequest {
  title: string;
  volume: CreateTrendRequestVolume;
  sentiment: string;
  active?: boolean;
}