# lib/api-zod/src/generated/types/clip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ClipPlatform } from "./clipPlatform";

export interface Clip {
  id: number;
  title: string;
  sourceUrl?: string;
  platform: ClipPlatform;
  timestamp?: string;
  keywords: string[];
  transcript?: string;
  season: string;
  notes?: string;
  createdAt: Date;
}