# lib/api-zod/src/generated/types/createClipRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateClipRequestPlatform } from "./createClipRequestPlatform";

export interface CreateClipRequest {
  title: string;
  sourceUrl?: string;
  platform: CreateClipRequestPlatform;
  timestamp?: string;
  keywords: string[];
  transcript?: string;
  season: string;
  notes?: string;
}