# lib/api-zod/src/generated/types/createEpisodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEpisodeRequestStatus } from "./createEpisodeRequestStatus";

export interface CreateEpisodeRequest {
  title: string;
  episodeNumber?: number;
  season: string;
  status?: CreateEpisodeRequestStatus;
  scheduledDate?: Date;
  topic?: string;
  notes?: string;
}