# lib/api-zod/src/generated/types/createTrendRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTrendRequestVolume } from "./createTrendRequestVolume";

export interface CreateTrendRequest {
  title: string;
  volume: CreateTrendRequestVolume;
  sentiment: string;
  active?: boolean;
}