# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeStatus } from "./episodeStatus";

export interface Episode {
  id: number;
  title: string;
  episodeNumber?: number;
  season: string;
  status: EpisodeStatus;
  scheduledDate?: Date;
  publishedDate?: Date;
  youtubeUrl?: string;
  topic?: string;
  notes?: string;
  createdAt: Date;
}