# lib/api-zod/src/generated/types/researchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchResultSourcesItem } from "./researchResultSourcesItem";

export interface ResearchResult {
  query: string;
  content: string;
  sources: ResearchResultSourcesItem[];
}