# lib/api-zod/src/generated/types/trend.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendVolume } from "./trendVolume";

export interface Trend {
  id: number;
  title: string;
  volume: TrendVolume;
  sentiment: string;
  active: boolean;
  createdAt: Date;
}