# lib/api-zod/src/generated/types/updateEpisodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMPYOG Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEpisodeRequestStatus } from "./updateEpisodeRequestStatus";

export interface UpdateEpisodeRequest {
  title?: string;
  episodeNumber?: number;
  season?: string;
  status?: UpdateEpisodeRequestStatus;
  scheduledDate?: Date;
  publishedDate?: Date;
  youtubeUrl?: string;
  topic?: string;
  notes?: string;
}